+++ .github/workflows/linter.yml
@@ -14,8 +14,8 @@ jobs:
     - name: Install Dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install -r requirements.txt
+        pip install -r src/requirements.txt
         pip install pylint
     - name: Run Pylint
       run: |
-        pylint *.py
+        pylint $(git ls-files '*.py')
