+++ install_all_requirements.sh
@@ -1,7 +1,7 @@
 #! /bin/bash
 
 # install dependences
-pip install -r requirements.txt
+pip install -r src/requirements.txt
 pip install -r emotional_speech_recognizing/requirements.txt
 pip install -r silero_demo/requirements.txt
