+++ src/modules/stt_module/audio_classes.py
