+++ src/modules/stt_module/voice_module.py
@@ -8,12 +8,12 @@ from telegram import Update
 from telegram.ext import CallbackContext
 from bson import json_util
 
-from whisper_module import get_att_whisper
-from audio_classes import RecognizedSentence
-from db import push_user_survey_progress, init_user, get_user_audio
+from src.modules.stt_module.whisper_module import get_att_whisper
+from src.modules.stt_module.audio_classes import RecognizedSentence
+from src.databases.db import push_user_survey_progress, init_user, get_user_audio
 
-from env_config import (DEBUG_MODE,
-                        DEBUG_ON, DEBUG_OFF)
+from src.env_config import (DEBUG_MODE,
+                            DEBUG_ON, DEBUG_OFF)
 
 
 def audio_to_text(filename):
