+++ src/modules/tts_module/silero_module.py
@@ -1,6 +1,6 @@
 import requests
 
-from env_config import SPEAKER, LANGUAGE, SAMPLE_RATE
+from src.env_config import SPEAKER, LANGUAGE, SAMPLE_RATE
 
 
 class VoiceSettings:
