+++ src/my_cron.py
@@ -7,8 +7,16 @@ from typing import List
 import schedule
 from telegram.ext import Updater
 
-from bot import ask_ready, resume_survey
-from db import get_schedule_list_for_feeling_ask, Schedule, get_users_not_finish_survey
+from src.commands.handlers import (
+    ask_ready,
+    resume_survey
+)
+
+from src.databases.db import (
+    get_schedule_list_for_feeling_ask,
+    Schedule,
+    get_users_not_finish_survey
+)
 
 
 MINUTES_FOR_LOOP = 1
