+++ src/utilities/translation.py
@@ -1,0 +1,14 @@
+import gettext
+
+from src.databases.db import (
+    get_user_pronoun
+)
+
+
+def set_translation(user):
+    pronoun = get_user_pronoun(user)
+    languages = 'ru_official' if pronoun else 'ru'
+
+    current_translation = gettext.translation('messages', localedir='locale', languages=[languages])
+    current_translation.install()
+    return current_translation
