+++ src/utilities/wrapper.py
@@ -1,9 +1,17 @@
 from telegram import Update
 
-from silero_module import bot_answer_audio, clear_audio_cache
+from src.modules.tts_module.silero_module import (
+    bot_answer_audio,
+    clear_audio_cache
+)
 
-from env_config import (DEBUG_MODE, DEBUG_ON, DEBUG_OFF)
-from db import get_user_mode
+from src.env_config import (
+    DEBUG_MODE,
+    DEBUG_ON,
+    DEBUG_OFF
+)
+
+from src.databases.db import get_user_mode
 
 
 def dialog(update: Update, text: str, reply_markup=None) -> None:
